requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
requests>=2.31
pydantic>=2.0
python-multipart>=0.0.9
jinja2>=3.1
assemblyai>=0.33.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiofiles>=23.2
httpx>=0.27.0
